Create Bulk Upload Google Drive
This project uses the Google Drive API in the Google Cloud Project platform in tandem with the PyDrive library to bulk upload files from a user's laptop to a select folder in the user's google drive account. The code is currently written to take user input, but the local file directory and Drive folder can be hardcoded if necessary so that no user input is needed.